+++ src/extension.ts
@@ -271,6 +271,7 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
                 },
                 extensionVariables: {
                     getOutputChannel: () => ext.outputChannel,
+                    getFocusedGroup: () => ext.focusedGroup,
                 },
                 testing: {
                     setOverrideAzureServiceFactory: (factory) => {